+++ client/src/components/App/App.tsx
@@ -2,28 +2,27 @@ import React from "react";
 import { connect } from "react-redux";
 import { ThemeProvider as EmotionThemeProvider } from "@emotion/react";
 import { StyledEngineProvider, ThemeProvider } from "@mui/material/styles";
-import { theme } from "./theme";
-import BottomBanner from "./BottomBanner/BottomBanner";
-
-import Controls from "./controls";
-import DatasetSelector from "./datasetSelector/datasetSelector";
-import Container from "./framework/container";
-import Layout from "./framework/layout";
-import LayoutSkeleton from "./framework/layoutSkeleton";
-import LeftSideBar from "./leftSidebar";
-import RightSideBar from "./rightSidebar";
-import Legend from "./continuousLegend";
-import MenuBar from "./menubar";
-import Header from "./NavBar";
-import actions from "../actions";
-import { RootState, AppDispatch } from "../reducers";
-import GlobalHotkeys from "./hotkeys";
-import { selectIsSeamlessEnabled } from "../selectors/datasetMetadata";
-import Graph from "./graph/graph";
-import DiffexNotice from "./diffexNotice";
-import Scatterplot from "./scatterplot/scatterplot";
-import PanelEmbedding from "./PanelEmbedding";
-import { BANNER_FEEDBACK_SURVEY_LINK } from "./BottomBanner/constants";
+import actions from "actions";
+import { RootState, AppDispatch } from "reducers";
+import Controls from "common/components/Controls/Controls";
+import { theme } from "util/theme";
+import BottomBanner from "../BottomBanner/BottomBanner";
+import DatasetSelector from "../DatasetSelector/DatasetSelector";
+import Container from "../framework/container";
+import Layout from "../framework/layout";
+import LayoutSkeleton from "../framework/layoutSkeleton";
+import LeftSideBar from "../leftSidebar";
+import RightSideBar from "../rightSidebar";
+import Legend from "../Legend/Legend";
+import MenuBar from "../menubar";
+import Header from "../NavBar";
+import GlobalHotkeys from "../hotkeys";
+import { selectIsSeamlessEnabled } from "../../selectors/datasetMetadata";
+import Graph from "../graph/graph";
+import DiffexNotice from "../DiffexNotice/DiffexNotice";
+import Scatterplot from "../scatterplot/scatterplot";
+import PanelEmbedding from "../PanelEmbedding";
+import { BANNER_FEEDBACK_SURVEY_LINK } from "../BottomBanner/constants";
 
 interface StateProps {
   loading: RootState["controls"]["loading"];
