+++ client/src/components/geneExpression/geneSet.tsx
@@ -3,8 +3,8 @@ import { FaChevronRight, FaChevronDown } from "react-icons/fa";
 import BrushableHistogram from "common/components/BrushableHistogram/BrushableHistogram";
 import { track } from "analytics";
 import { EVENTS } from "analytics/events";
+import Truncate from "common/components/Truncate/Truncate";
 import Gene from "./gene";
-import Truncate from "../util/truncate";
 import * as globals from "~/globals";
 import GenesetMenus from "./menus/genesetMenus";
 import EditGenesetNameDialogue from "./menus/editGenesetNameDialogue";
