+++ client/src/util/transientLocalStorage.ts
